The changes made to the code involve renaming multiple files and updating their contents to reflect the new structure. Specifically, a new folder named 'temperature' has been created within the 'src/command/impl/lr322' directory. The following files were moved into this new folder:
1. CoolantTempCommandImpl.cpp
2. CoolantTempCommandImpl.h
3. ComprTempCommandImpl.cpp
4. ComprTempCommandImpl.h
5. GbTempCommandImpl.cpp
6. GbTempCommandImpl.h
7. OilTempCommandImpl.cpp
8. OilTempCommandImpl.h
9. AmbientAirTempCommandImpl.cpp
10. AmbientAirTempCommandImpl.h
11. IntakeAirTempCommandImpl.cpp
12. IntakeAirTempCommandImpl.h
13. RDTempCommandImpl.cpp
14. RDTempCommandImpl.h

In addition to these file movements, the 'CoolantTempCommandImpl.h' file was updated to include necessary declarations for a new variable 'coolantTemp'. This change suggests that the coolant temperature will now be tracked or calculated by this class.

Overall, these modifications appear to organize the files related to temperature-related commands more efficiently beneath a dedicated 'temperature' subdirectory and potentially prepare for future enhancements or refinements in handling temperature data.